revised net engine

+++ src/net/server.h
@@ -15,6 +15,7 @@
 
 #include "../common.h"
 #include "client.h"
+#include "protocol.h"
 
 #define ADDR(a, b, c, d) ( a << 24 ) | ( b << 16 ) | ( c << 8 ) | d;
 
@@ -32,7 +33,7 @@ struct data {
 int server_createUDPSocket(unsigned int port);
 int server_receive_packets();
 
-int server_start(unsigned int port);
+int server_start(unsigned short int port);
 int server_send_packet(unsigned char *data, size_t len, struct client* c);
 int server_remove_client(int id);
 struct client *get_client_by_id(int id);
